Removed Duplicate Files and Minor UI Changes - Curved Barchart without x and y-axis grid - Balanced Text Spacing - Added carousel type scrolling in view options - watchOS App will now share the same model and viewmodel file as the iOS App

+++ Vertanom/Model/Device.swift
@@ -16,7 +16,7 @@ enum DeviceStatus: String, Codable {
 
 struct SensorValue: Identifiable, Codable, Hashable {
     let date: Date
-    var value: Int
+    var value: Double
     var id: Date { date }
 }
